DispCal: Show menu only if feature supported on device
We have a new kernel sysfs file that tells us whether the device
is compatible or not with DispCal, or if the panel found in the
device has been calibrated.

Check this sysfs file and hide the DispCal menu if the feature
is not declared as available by the kernel.